fix(taxonomy): exercise real flag parser in resolve tests
Extract the flag-stripping logic from cmdTaxonomy's resolve case into a
standalone stripResolveFlags() function in cli-args.ts. Both the CLI and
the test file now import the same function, so tests break if the real
parser regresses.

Addresses unresolved P1 finding from PR #433 (thread PRRT_kwDORJXyws57s46h).
