+++ tests/taxonomy-cli-boolean-flag.test.ts
@@ -3,78 +3,59 @@
  * text argument in the `taxonomy resolve` flag-stripping loop.
  *
  * Also tests the generic coerceBool helper (Finding 3).
+ *
+ * Exercises the REAL stripResolveFlags implementation from cli-args.ts
+ * (not a local copy) so regressions in the CLI are caught by these tests.
  */
 
 import { describe, it } from "node:test";
 import assert from "node:assert/strict";
 import { coerceBool, coerceInstallExtension } from "../packages/remnic-core/src/connectors/coerce.js";
-
-// ── Boolean flag stripping logic (Finding 1) ───────────────────────────────
-//
-// The resolve case strips --flag <value> pairs before joining remaining tokens
-// as the input text. Boolean flags like --json have no trailing value, so the
-// loop must not skip the token after them.
-//
-// We replicate the exact stripping logic here so the test stays self-contained
-// and doesn't need to import the full CLI entry.
-
-const BOOLEAN_FLAGS = new Set(["--json"]);
-
-function stripFlags(args: string[]): string[] {
-  const textParts: string[] = [];
-  for (let i = 0; i < args.length; i++) {
-    if (args[i].startsWith("--")) {
-      if (!BOOLEAN_FLAGS.has(args[i])) {
-        // Key-value flag: skip the flag and its value (next token)
-        i++;
-      }
-      continue;
-    }
-    textParts.push(args[i]);
-  }
-  return textParts;
-}
+import {
+  stripResolveFlags,
+  TAXONOMY_RESOLVE_BOOLEAN_FLAGS,
+} from "../packages/remnic-cli/src/cli-args.js";
 
 describe("taxonomy resolve flag stripping", () => {
   it("--json does not eat the following text argument", () => {
     const args = ["--json", "hello", "world"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, ["hello", "world"]);
   });
 
   it("--category (key-value) correctly skips the value", () => {
     const args = ["--category", "preference", "my", "text"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, ["my", "text"]);
   });
 
   it("--json before --category works correctly", () => {
     const args = ["--json", "--category", "fact", "some", "input"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, ["some", "input"]);
   });
 
   it("--category before --json works correctly", () => {
     const args = ["--category", "fact", "--json", "some", "input"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, ["some", "input"]);
   });
 
   it("--json at the end does not cause out-of-bounds", () => {
     const args = ["hello", "--json"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, ["hello"]);
   });
 
   it("no flags returns all tokens", () => {
     const args = ["hello", "world"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, ["hello", "world"]);
   });
 
   it("only --json returns empty", () => {
     const args = ["--json"];
-    const result = stripFlags(args);
+    const result = stripResolveFlags(args, TAXONOMY_RESOLVE_BOOLEAN_FLAGS);
     assert.deepStrictEqual(result, []);
   });
 });
